Make Tricky68k.app depend on the linker from binutils, and build just the linker, not all the binutils+gdb.

+++ binutils/buildscript.sh
@@ -1,30 +1,36 @@
 #!/bin/bash
 
-set -e
+set -ek
+DESTDIR="$DSTROOT"
+export DESTDIR
+env
 
 mkdir -p "$CONFIGURATION_TEMP_DIR"
+mkdir -p "$CONFIGURATION_BUILD_DIR"
 cd "$CONFIGURATION_TEMP_DIR"
 
 ln -fs "$SRCROOT/binutils-gdb/" "$PROJECT_TEMP_DIR"
 
-case "$ACTION" in
-  clean)
+if [ "$ACTION" == "clean" ]; then
     if [ -e Makefile ]; then
       make clean;
       rm -rf ./*
     fi
 # This uninstall presumes that BUILT_PRODUCTS_DIR is install location for
 # binutils only!
-    rm -rf "$BUILT_PRODUCTS_DIR/bin"
-    rm -rf "$BUILT_PRODUCTS_DIR/include"
-    rm -rf "$BUILT_PRODUCTS_DIR/m68k-elf"
-    rm -rf "$BUILT_PRODUCTS_DIR/share"
+    rm -rf "$CONFIGURATION_BUILD_DIR/bin"
+    rm -rf "$CONFIGURATION_BUILD_DIR/include"
+    rm -rf "$CONFIGURATION_BUILD_DIR/m68k-elf"
+    rm -rf "$CONFIGURATION_BUILD_DIR/share"
+    rm -f "$CONFIGURATION_BUILD_DIR/m68k-elf-ld"
     exit 0
-    ;;
-esac
+fi
 
 if [ ! -e Makefile ]; then
   "$PROJECT_TEMP_DIR/binutils-gdb/configure" $@;
 fi
-make
-make install
+make "all-$TARGET_NAME"
+if [ "$ACTION" == "install" ]; then
+  make "install-$TARGET_NAME"
+fi
+cp "$CONFIGURATION_TEMP_DIR/ld/ld-new" "$CONFIGURATION_BUILD_DIR/m68k-elf-ld"
